Thrash Master 2000
I truly have little idea what happened on this commit. I have to get
better about not thrashing. I spent most of today just trying to get
the front end to create a record for pictures with the recipe id
associated. It does that now, and displays the most recent entry to
the picture component. However, if there are multiple entries then
you only get the first of them when things load.

Need to handle entry uniqueness. And make it so I can edit and update
at the same time. This may need to be it's own component.